pip-requirements.txt
pika>=1.1.0
redis==2.10.1
requests==2.20.0
pyOpenSSL==18.0.0
ckantoolkit==0.0.3
